Add sqlite-functions crate
The idea is to have helper SQLite functions injected into every
connections that can be used by API queries and subscriptions.

The first one is 'corro_json_contains' which returns true if the
first argument interpreted as JSON object is fully contained withing
the second argument interpreted as JSON object.