Work around channel overflow for freshly created subscriptions (#184)
When a new subscription is created, the data is sent directly to the
broadcast channel. If this channel overflows because the receiver is
not fast enough, the error is not detected by `select!` and the client
is left hanging forever.

Fix the error detection, so that the client can actually re-subscribe.
Since it re-subscribes with existing subscription, the initial data
is sent to the mpsc channel that doesn't suffer from the overflow
issue.