corro-client: fix subscription io::Error backoff
Make sure the count is only cleared once the client get a successful
request or encounters a non-IO error. Otherwise it never finishes.

Also, since the upper layer usually retries on its own, limit the numer
of retries to just 10.

This should be enough in case of a simple server restart, but will
provide faster reaction times if the host is down completely.