Increase WAL checkpoint timeout
busy_handler of checkpoints will wait until all readers are on the same
latest snapshot, NOT when all readers complete. With huge subs we often
have readers that take very long to complete, and in that case we might
keep missing checkpoints until the WAL blows up in size, slowing down
readers even further.

Because we don't actually need to wait for all readers to fully
complete, and it doesn't block readers, try increasing the timeout by
quite a lot.